+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+- Augmenting the `ModelEvents` interface it is now possible to extend model events with custom events.
+
+### Deprecated
+
+- `ModelEvent` and `ModelEventValue` have been deprecated in favour of the `ModelEvents` interface.
+
 ## [v0.5.2](https://github.com/NoelDeMartin/soukai/releases/tag/v0.5.2) - 2023-12-17
 
 ### Added